% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{anjouan}
\alias{anjouan}
\title{Anjouan
Draw a map for Anjouan Island}
\usage{
anjouan(x = "anjouan", pref = FALSE)
}
\arguments{
\item{x}{Name of the data set to use. The default is \code{anjouan()}, It draws Anjouan Island as one object without prefecture.}

\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}
}
\value{
The data set used is in \code{sf} format
}
\description{
Anjouan
Draw a map for Anjouan Island
}
\examples{
## Map Anjouan Island.
anjouan ()
## Map Anjouan Island with prefecture area.
anjouan (pref = TRUE)


}
