% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{comorosmap}
\alias{comorosmap}
\title{Comoro Islands Maps}
\usage{
comorosmap(x = "country", region = FALSE)
}
\arguments{
\item{x}{name of the data set to use, the default is comorompaps(x = "country",region = False)}

\item{region}{binary variable: TRUE or FALSE. for each x, choose if region = TRUE or FALSE}
}
\value{
the data set used, in \code{sf} format
}
\description{
Draw a map of Comoros Islands: by Country, by Island, by Prefecture and by Commune
}
\details{
\code{comoromaps()}uses the sf package to plot by ploting the geometry rather than every column and leave the plot region ready for overplotting with other data
}
