% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{grandeComore}
\alias{grandeComore}
\title{Draw a map for Grande Comore Island}
\usage{
grandeComore(x = "grande comore", pref = FALSE)
}
\arguments{
\item{x}{Name of the data set to use. The default is \code{grandeComore()}, It draws Grande Comore Island as one object without prefecture.}

\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw a map for Grande Comore Island
}
\examples{
grandeComore ()
grandeComore (pref = TRUE)

}
