% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{moheli}
\alias{moheli}
\title{Draw a map for Moheli Islands}
\usage{
moheli(x = "moheli", pref = FALSE)
}
\arguments{
\item{x}{Name of the data set to use. The default is \code{moheli()}, It draws Moheli Islands as one object without prefecture.}

\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw a map for Moheli Islands
}
\examples{
moheli ()
moheli (pref = TRUE)

}
