% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{comorosmap}
\alias{comorosmap}
\title{Comoro Islands Map}
\usage{
comorosmap(x = "country", pref = FALSE)
}
\arguments{
\item{x}{Name of the data set to use. The default is \code{comorosmpap()}, It draws Comoro Islands as one object without commune.}

\item{pref}{Choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw  maps of Comoros Islands.
}
\details{
\code{comorosmap()}uses the sf package to plot by plotting the geometry rather than every column and leave the plot pref ready for overplotting with other data
}
\examples{
comorosmap() ## Map Comoro Islands as one object without prefecture area
comorosmap("country",pref=TRUE) ## Map Comoro Islands as one object with prefecture area
comorosmap("island",pref=FALSE) ## Map Comoros Islands as 3 object (Grande Comore, Anjouan, Mohéli) without prefecture area.
comorosmap("island",pref=TRUE) ## Map Comoros Islands as 3 object (Grande Comore, Anjouan, Mohéli) with prefecture area.
comorosmap("anjouan",pref=FALSE) ## Map Anjouan island without prefcture area
comorosmap("anjouan",pref=TRUE) ## Map Anjouan island with prefcture area
comorosmap("moheli",pref=FALSE) ## Map Mohéli island without prefcture area
comorosmap("moheli",pref=TRUE) ## Map Mohéli island with prefcture area
comorosmap("grande comore",pref=FALSE) ## Map Grande Comore island without prefcture area
comorosmap("grande comore",pref=TRUE) ## Map Grande Comore island with prefcture area


}
