% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps.R
\name{comorosmap}
\alias{comorosmap}
\title{Comoro Islands Map}
\usage{
comorosmap(x = "country", pref = FALSE)
}
\arguments{
\item{x}{Name of the data set to use. The default is \code{comorosmpap()}, It draws Comoro Islands as one object without commune.}

\item{pref}{TRUE or FALSE. For each x, choose to map with prefecture area ("pref" = TRUE) or without prefecture area ("pref" = FALSE)}
}
\value{
The data set used is in \code{sf} format
}
\description{
Draw  maps of Comoros Islands.  by Country, by Island, by Prefecture and by Commune
}
\details{
\code{comorosmaps()}uses the sf package to plot by plotting the geometry rather than every column and leave the plot pref ready for overplotting with other data
}
\examples{
comorosmap() Map Comoro Islands as one object without prefecture area
comorosmap(region=TRUE) Map Comoro Islands as one object with prefecture area
comorosmap("isaland",region=FALSE) Map Comoros Islands as 3 object (Grande Comore, Anjouan, Mohéli) without prefecture area.
comorosmap("isaland",region=TRUE) Map Comoros Islands as 3 object (Grande Comore, Anjouan, Mohéli) with prefecture area.
Comorosmap("anjouan",region=FALSE) Map Anjouan island without prefcture area
Comorosmap("anjouan",region=TRUE) Map Anjouan island with prefcture area
Comorosmap("mohéli",region=FALSE) Map Mohéli island without prefcture area
Comorosmap("mohéli",region=TRUE) Map Mohéli island with prefcture area
Comorosmap("grande comore",region=FALSE) Map Grande Comore island without prefcture area
Comorosmap("grande comore",region=TRUE) Map Grande Comore island with prefcture area


}
