% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps_package.R
\docType{package}
\name{comorosmaps}
\alias{comorosmaps}
\title{comorosmaps: Comoro Islands Maps}
\description{
A package for mapping Comoro Islands costaline. The package provides three main functions:\code{comorosmap()},\code{grandeComore()}, \code{anjouan()}, \code{moheli()}
}
\section{comorosmap function}{

The comorosmap function allows to plot 3 main maps:
\itemize{
\item country: Comoro Islands map as a single object
\item island: Comoro Islands map as triple objects. Each object represents an  Island.
\item "grande comore","moheli", "anjouan": Comoro Islands map island by island
}

For each map, comorosmap function allows to map with or without administrative region.
}

\seealso{
Useful links:
\itemize{
\item https://github.com/hhousni/comorosmaps
\item Report bugs at https://github.com/hhousni/comorosmaps/issues
}
}
\author{
Housni Hassani happrend2@gmail.com
}
