% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorosmaps_package.R
\docType{package}
\name{comorosmaps}
\alias{comorosmaps}
\title{comorosmaps: Comoro Islands Maps}
\description{
A package for mapping Comoro Islands coastline and prefecture area. Data can be drawn or accessed directly as simple features objects.
The package includes simple function for each Island: \code{grandeComore()}, \code{anjouan()}, \code{moheli()} and a function for the country \code{comoros()}
}
\seealso{
Useful links:
\itemize{
\item https://github.com/hhousni/comorosmaps
\item Report bugs at https://github.com/hhousni/comorosmaps/issues
}
}
\author{
Housni Hassani happrend2@gmail.com
}
